Move decider_client_from_config to reddit_experiments
