+++ reddit_decider/__init__.py
@@ -223,7 +223,7 @@ class DeciderContextFactory(ContextFactory):
         event_logger: Optional[EventLogger] = None,
         timeout: Optional[float] = None,
         backoff: Optional[float] = None,
-        request_field_extractor: Callable[[BaseplateRequest], Dict[str, str]] = None
+        request_field_extractor: Optional[Callable[[BaseplateRequest], Dict[str, str]]] = None
     ):
         self._filewatcher = FileWatcher(path=path, parser=init_decider_parser, timeout=timeout, backoff=backoff)
         self._event_logger = event_logger
@@ -283,69 +283,3 @@ class DeciderContextFactory(ContextFactory):
             context_name=name,
             event_logger=self._event_logger,
         )
-
-
-def decider_client_from_config(
-    app_config: config.RawConfig,
-    event_logger: EventLogger,
-    prefix: str = "experiments.",
-    request_field_extractor: Callable[[BaseplateRequest], Dict[str, str]] = None
-) -> DeciderContextFactory:
-    """Configure and return an :py:class:`DeciderContextFactory` object.
-
-    The keys useful to :py:func:`decider_client_from_config` should be prefixed, e.g.
-    ``experiments.path``, etc.
-
-    Supported keys:
-
-    ``path`` (optional)
-        The path to the experiment configuration file generated by the
-        experiment configuration fetcher daemon.
-    ``timeout`` (optional)
-        The time that we should wait for the file specified by ``path`` to
-        exist.  Defaults to `None` which is `infinite`.
-    ``backoff`` (optional)
-        The base amount of time for exponential backoff when trying to find the
-        experiments config file. Defaults to no backoff between tries.
-    ``request_field_extractor`` (optional) used to populate
-        "app_name" & "build_number" fields in DeciderContext()
-
-    :param raw_config: The application configuration which should have
-        settings for the experiments client.
-    :param event_logger: The EventLogger to be used to log bucketing events.
-    :param prefix: the prefix used to filter keys (defaults to "experiments.").
-
-    """
-    assert prefix.endswith(".")
-    config_prefix = prefix[:-1]
-
-    cfg = config.parse_config(
-        app_config,
-        {
-            config_prefix: {
-                "path": config.Optional(config.String, default="/var/local/experiments.json"),
-                "timeout": config.Optional(config.Timespan),
-                "backoff": config.Optional(config.Timespan),
-            }
-        },
-    )
-    options = getattr(cfg, config_prefix)
-
-    # pylint: disable=maybe-no-member
-    if options.timeout:
-        timeout = options.timeout.total_seconds()
-    else:
-        timeout = None
-
-    if options.backoff:
-        backoff = options.backoff.total_seconds()
-    else:
-        backoff = None
-
-    return DeciderContextFactory(
-        path=options.path,
-        event_logger=event_logger,
-        timeout=timeout,
-        backoff=backoff,
-        request_field_extractor=request_field_extractor
-    )
