+++ reddit_experiments/__init__.py
@@ -3,6 +3,7 @@ import logging
 import warnings
 
 from enum import Enum
+from typing import Callable
 from typing import Dict
 from typing import Optional
 from typing import overload
@@ -11,6 +12,7 @@ from typing import Set
 
 from baseplate import Span
 from baseplate.clients import ContextFactory
+from baseplate.frameworks.pyramid import BaseplateRequest
 from baseplate.lib import config
 from baseplate.lib import warn_deprecated
 from baseplate.lib.events import DebugLogger
@@ -19,6 +21,7 @@ from baseplate.lib.file_watcher import FileWatcher
 from baseplate.lib.file_watcher import WatchedFileNotAvailableError
 from reddit_edgecontext import User
 
+from reddit_decider import DeciderContextFactory
 from reddit_experiments.providers import parse_experiment
 from reddit_experiments.providers.base import Experiment
 
@@ -396,3 +399,69 @@ def experiments_client_from_config(
         backoff = None
 
     return ExperimentsContextFactory(options.path, event_logger, timeout=timeout, backoff=backoff)
+
+
+def decider_client_from_config(
+    app_config: config.RawConfig,
+    event_logger: EventLogger,
+    prefix: str = "experiments.",
+    request_field_extractor: Optional[Callable[[BaseplateRequest], Dict[str, str]]] = None,
+) -> DeciderContextFactory:
+    """Configure and return an :py:class:`DeciderContextFactory` object.
+
+    The keys useful to :py:func:`decider_client_from_config` should be prefixed, e.g.
+    ``experiments.path``, etc.
+
+    Supported keys:
+
+    ``path`` (optional)
+        The path to the experiment configuration file generated by the
+        experiment configuration fetcher daemon.
+    ``timeout`` (optional)
+        The time that we should wait for the file specified by ``path`` to
+        exist.  Defaults to `None` which is `infinite`.
+    ``backoff`` (optional)
+        The base amount of time for exponential backoff when trying to find the
+        experiments config file. Defaults to no backoff between tries.
+    ``request_field_extractor`` (optional) used to populate
+        "app_name" & "build_number" fields in DeciderContext()
+
+    :param raw_config: The application configuration which should have
+        settings for the experiments client.
+    :param event_logger: The EventLogger to be used to log bucketing events.
+    :param prefix: the prefix used to filter keys (defaults to "experiments.").
+
+    """
+    assert prefix.endswith(".")
+    config_prefix = prefix[:-1]
+
+    cfg = config.parse_config(
+        app_config,
+        {
+            config_prefix: {
+                "path": config.Optional(config.String, default="/var/local/experiments.json"),
+                "timeout": config.Optional(config.Timespan),
+                "backoff": config.Optional(config.Timespan),
+            }
+        },
+    )
+    options = getattr(cfg, config_prefix)
+
+    # pylint: disable=maybe-no-member
+    if options.timeout:
+        timeout = options.timeout.total_seconds()
+    else:
+        timeout = None
+
+    if options.backoff:
+        backoff = options.backoff.total_seconds()
+    else:
+        backoff = None
+
+    return DeciderContextFactory(
+        path=options.path,
+        event_logger=event_logger,
+        timeout=timeout,
+        backoff=backoff,
+        request_field_extractor=request_field_extractor,
+    )
